refactor : 정적 유효성 검사 코드 위치 변경
