+++ src/constants/projectConstants.ts
@@ -1,5 +1,7 @@
 //constants
 
+import { z } from 'zod';
+
 export const PROJECT_DATA = [
   {
     id: '1',
@@ -73,3 +75,86 @@ export const CAREER_INPUT = [
     type: 'text',
   },
 ];
+
+export const createProjectScheme = z.object({
+  startDate: z
+    .string({ required_error: '시작 날짜를 입력해주세요.' })
+    .refine((date) => !isNaN(Date.parse(date)), {
+      message: '유효한 날짜를 입력해주세요.',
+    }),
+  endDate: z
+    .string({ required_error: '종료 날짜를 입력해주세요.' })
+    .refine((date) => !isNaN(Date.parse(date)), {
+      message: '유효한 날짜를 입력해주세요.',
+    }),
+
+  title: z
+    .string({ message: '프로젝트 제목을 입력해주세요.' })
+    .min(1, { message: '프로젝트 제목을 입력해주세요.' }),
+
+  maxVolunteers: z.coerce
+    .number({ message: '모집 인원을 입력해주세요.' })
+    .min(1, { message: '모집 인원은 1명 이상이어야 합니다.' })
+    .max(1000, { message: '모집 인원은 1000명 이하이어야 합니다.' }),
+  startDatePre: z
+    .string({ required_error: '종료 날짜를 입력해주세요.' })
+    .refine((date) => !isNaN(Date.parse(date)), {
+      message: '유효한 날짜를 입력해주세요.',
+    }),
+  field: z.number({ message: '진행 방식을 선택 해주세요.' }),
+  duration: z.coerce
+    .number({ message: '예상 기간을 입력해주세요.' })
+    .positive({ message: '예상 기간은 1 이상이어야 합니다.' })
+    .max(365, { message: '예상 기간은 365일을 초과할 수 없습니다.' }),
+  position: z
+    .array(z.number({ message: '숫자로 입력 되어야 합니다.' }))
+    .min(1, { message: '1개의 분야를 선택해주세요.' }),
+  newBy: z.boolean().optional(),
+  languages: z
+    .array(z.number({ message: '숫자로 입력 되어야 합니다.' }))
+    .min(1, { message: '최소 1개 이상의 언어를 선택해주세요.' }),
+
+  markdownEditor: z
+    .string({ message: '프로젝트 내용을 입력해주세요.' })
+    .min(10, { message: '프로젝트 내용은 최소 10자 이상이어야 합니다.' }),
+});
+
+export const ApplyScheme = z.object({
+  email: z
+    .string()
+    .nonempty({ message: '이메일을 입력해주세요.' })
+    .email({ message: '이메일 형식으로 입력해주세요.' }),
+  phone: z
+    .string({ message: '전화번호를 입력해주세요.' })
+    .array()
+    .nonempty({ message: '전화번호를 입력해주세요.' }),
+  wantToSay: z.string().optional(),
+  careers: z
+    .array(
+      z
+        .object({
+          name: z.string().nonempty({ message: '경력명을 입력해주세요.' }),
+          periodStart: z
+            .string()
+            .nonempty({ message: '시작 날짜를 입력해주세요.' })
+            .refine((date) => !isNaN(Date.parse(date)), {
+              message: '유효한 날짜를 입력해주세요.',
+            }),
+          periodEnd: z
+            .string()
+            .nonempty({ message: '종료 날짜를 입력해주세요.' })
+            .refine((date) => !isNaN(Date.parse(date)), {
+              message: '유효한 날짜를 입력해주세요.',
+            }),
+          role: z.string().nonempty({ message: '역할을 입력해주세요.' }),
+        })
+        .refine(
+          (data) => new Date(data.periodStart) < new Date(data.periodEnd),
+          {
+            message: '시작 날짜는 종료 날짜보다 이전이어야 합니다.',
+            path: ['periodStart'],
+          }
+        )
+    )
+    .optional(),
+});
