+++ src/models/joinProject.ts
@@ -1,3 +1,8 @@
+import { z } from 'zod';
+import { ApplyScheme } from '../constants/projectConstants';
+
+export type ApplySchemeType = z.infer<typeof ApplyScheme>;
+
 export interface Career {
   name: string;
   periodStart: string;
