+++ src/pages/apply/Apply.tsx
@@ -2,10 +2,9 @@ import * as S from './Apply.styled';
 import Input from '../../components/projectFormComponents/inputComponent/InputComponent';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { z } from 'zod';
 import { useParams } from 'react-router-dom';
 import { formatDate } from '../../util/format';
-import { joinProject } from '../../models/joinProject';
+import { ApplySchemeType, joinProject } from '../../models/joinProject';
 import CareersComponent from '../../components/applyComponents/careersComponent/CareersComponent';
 import PhoneComponent from '../../components/applyComponents/phoneComponent/PhoneComponent';
 import LoadingSpinner from '../../components/common/loadingSpinner/LoadingSpinner';
@@ -15,48 +14,7 @@ import useAuthStore from '../../store/authStore';
 import { useEffect } from 'react';
 import useGetProjectData from '../../hooks/useGetProjectData';
 import useApplyProject from '../../hooks/ProjectHooks/useApplyProject';
-
-const ApplyScheme = z.object({
-  email: z
-    .string()
-    .nonempty({ message: '이메일을 입력해주세요.' })
-    .email({ message: '이메일 형식으로 입력해주세요.' }),
-  phone: z
-    .string({ message: '전화번호를 입력해주세요.' })
-    .array()
-    .nonempty({ message: '전화번호를 입력해주세요.' }),
-  wantToSay: z.string().optional(),
-  careers: z
-    .array(
-      z
-        .object({
-          name: z.string().nonempty({ message: '경력명을 입력해주세요.' }),
-          periodStart: z
-            .string()
-            .nonempty({ message: '시작 날짜를 입력해주세요.' })
-            .refine((date) => !isNaN(Date.parse(date)), {
-              message: '유효한 날짜를 입력해주세요.',
-            }),
-          periodEnd: z
-            .string()
-            .nonempty({ message: '종료 날짜를 입력해주세요.' })
-            .refine((date) => !isNaN(Date.parse(date)), {
-              message: '유효한 날짜를 입력해주세요.',
-            }),
-          role: z.string().nonempty({ message: '역할을 입력해주세요.' }),
-        })
-        .refine(
-          (data) => new Date(data.periodStart) < new Date(data.periodEnd),
-          {
-            message: '시작 날짜는 종료 날짜보다 이전이어야 합니다.',
-            path: ['periodStart'],
-          }
-        )
-    )
-    .optional(),
-});
-
-export type ApplySchemeType = z.infer<typeof ApplyScheme>;
+import { ApplyScheme } from '../../constants/projectConstants';
 
 const Apply = () => {
   const { projectId } = useParams();
