+++ src/pages/createProject/CreateProject.tsx
@@ -8,51 +8,8 @@ import ProjectInformationInput from '../../components/projectFormComponents/proj
 import Modal from '../../components/common/modal/Modal';
 import { useModal } from '../../hooks/useModal';
 import LoadingSpinner from '../../components/common/loadingSpinner/LoadingSpinner';
-import useAuthStore from '../../store/authStore';
 import useCreateProject from '../../hooks/ProjectHooks/useCreateProject';
-
-export const createProjectScheme = z.object({
-  startDate: z
-    .string({ required_error: '시작 날짜를 입력해주세요.' })
-    .refine((date) => !isNaN(Date.parse(date)), {
-      message: '유효한 날짜를 입력해주세요.',
-    }),
-  endDate: z
-    .string({ required_error: '종료 날짜를 입력해주세요.' })
-    .refine((date) => !isNaN(Date.parse(date)), {
-      message: '유효한 날짜를 입력해주세요.',
-    }),
-
-  title: z
-    .string({ message: '프로젝트 제목을 입력해주세요.' })
-    .min(1, { message: '프로젝트 제목을 입력해주세요.' }),
-
-  maxVolunteers: z.coerce
-    .number({ message: '모집 인원을 입력해주세요.' })
-    .min(1, { message: '모집 인원은 1명 이상이어야 합니다.' })
-    .max(1000, { message: '모집 인원은 1000명 이하이어야 합니다.' }),
-  startDatePre: z
-    .string({ required_error: '종료 날짜를 입력해주세요.' })
-    .refine((date) => !isNaN(Date.parse(date)), {
-      message: '유효한 날짜를 입력해주세요.',
-    }),
-  field: z.number({ message: '진행 방식을 선택 해주세요.' }),
-  duration: z.coerce
-    .number({ message: '예상 기간을 입력해주세요.' })
-    .positive({ message: '예상 기간은 1 이상이어야 합니다.' })
-    .max(365, { message: '예상 기간은 365일을 초과할 수 없습니다.' }),
-  position: z
-    .array(z.number({ message: '숫자로 입력 되어야 합니다.' }))
-    .min(1, { message: '1개의 분야를 선택해주세요.' }),
-  newBy: z.boolean().optional(),
-  languages: z
-    .array(z.number({ message: '숫자로 입력 되어야 합니다.' }))
-    .min(1, { message: '최소 1개 이상의 언어를 선택해주세요.' }),
-
-  markdownEditor: z
-    .string({ message: '프로젝트 내용을 입력해주세요.' })
-    .min(10, { message: '프로젝트 내용은 최소 10자 이상이어야 합니다.' }),
-});
+import { createProjectScheme } from '../../constants/projectConstants';
 
 const CreateProject = () => {
   const {
@@ -79,14 +36,6 @@ const CreateProject = () => {
   const { createProject, isLoading } = useCreateProject({
     handleModalOpen,
   });
-  const userId = useAuthStore((state) => state.userData?.id);
-  if (!userId) {
-    return (
-      <Modal isOpen={isOpen} onClose={handleModalClose}>
-        {message}
-      </Modal>
-    );
-  }
 
   const handleSubmit = (data: z.infer<typeof createProjectScheme>) => {
     const formData: FormData = {
